- JSoundEditor update. (you can open and save your own files yet) - Bugfixes - StreamUtils class for better looking code ;) - AudioFormat16Filter can convert a SoundInput in another AudioFormat.   (it is called AudioFormat16Filter because it just works with 16bit samples)
ZeroLuck

git-svn-id: http://svn.code.sf.net/p/jblocks/code-0/trunk@70 9daefd43-015a-43da-9b33-cb4dc19dcf47

+++ org/jblocks/soundeditor/Track.java
@@ -33,7 +33,7 @@ class Track {
         this.name = name;
         this.spp = samplesPerPixel;
     }
-    
+
     public String getName() {
         return name;
     }
@@ -64,11 +64,10 @@ class Track {
      */
     public short[] createUserView() {
         int length = end - offset;
-        short[] view = new short[length / spp];
+        short[] view = new short[(int) ((double) length / spp)];
         for (int i = 0; i < view.length; i++) {
             view[i] = samples[offset + i * spp];
         }
         return view;
     }
-
 }
